pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stopmo-xcode"
version = "0.1.0"
description = "Deterministic stop-motion RAW to LogC3/AWG transcoder"
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "stopmo-xcode" }]
license = { text = "MIT" }
dependencies = [
  "PyYAML>=6.0",
  "numpy>=1.24",
]

[project.optional-dependencies]
watch = ["watchdog>=4.0"]
raw = ["rawpy>=0.20"]
ocio = ["opencolorio>=2.3"]
io = ["tifffile>=2024.2"]
video = []
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[project.scripts]
stopmo-xcode = "stopmo_xcode.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
